# Approximate object references for ReferenceCountMap