ReferenceCountMap maps from objects to reference counts. Its keys are
stored in a single object array. In some extreme cases, the array grows
to hundreds of millions of objects. This can cause G1GC to push huge
amounts of object pointers to a task queue, which can both lead to GC
overhead as well as queue overflow. The problem can be solved by using
big arrays to store objects. However, the performance overhead (other than
GC) can still be high. This patch combines both identity hash code and
the size of the objects as the key to identity an object. Collision can
happen but according to benchmark is neglectable.