@losipiuk 2018-01-30
What does "approximate" in commit message relate to?