@highker 2018-01-30
@losipiuk the approach in this PR does not guarantee a precise memory tracking. We may ignore hash collisions.