@losipiuk 2018-01-30
> @losipiuk the approach in this PR does not guarantee a precise memory tracking. We may ignore hash collisions.

Yeah, right. I was living in a world where one can assume `identityHashMap` guarantees no collisions. Learned new thing today :)