@wenleix 2018-01-31
@highker : Can you paste the distribution of duplicate identity hash code in benchmark?