@highker 2018-01-31
collision out of 100 million objeccts:
```
1:	95456024
2:	2220373
3:	33866
4:	403
5:	4