@highker 2018-02-07
Comments address; thanks for review