@highker 2018-02-07
@martint, benchmark updated:

With this patch:
```
Benchmark                                                                     (arrayType)   Mode  Cnt        Score        Error   Units
BenchmarkReferenceCountMap.benchmarkInserts                                           int  thrpt   40  5886062.220 ± 309846.728   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                            int  thrpt   40      309.965 ±     15.293  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                       int  thrpt   40       55.464 ±      1.646    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                   int  thrpt   40       18.320 ±     22.365  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm              int  thrpt   40        3.779 ±      4.913    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                      int  thrpt   40      281.176 ±    288.476  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm                 int  thrpt   40       53.629 ±     54.668    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space               int  thrpt   40        7.884 ±     15.072  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space.norm          int  thrpt   40        1.583 ±      2.909    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                                 int  thrpt   40       11.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                                  int  thrpt   40      742.000                   ms
BenchmarkReferenceCountMap.benchmarkInserts                                        double  thrpt   40  6451697.048 ± 378030.818   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                         double  thrpt   40      337.807 ±     19.492  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                    double  thrpt   40       55.313 ±      1.368    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                double  thrpt   40       27.440 ±     31.566  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm           double  thrpt   40        5.039 ±      6.069    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                   double  thrpt   40      343.109 ±    334.970  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm              double  thrpt   40       60.003 ±     57.603    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space            double  thrpt   40       12.746 ±     22.013  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space.norm       double  thrpt   40        2.502 ±      4.317    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                              double  thrpt   40       12.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                               double  thrpt   40     1153.000                   ms
BenchmarkReferenceCountMap.benchmarkInserts                                          long  thrpt   40  5638105.996 ± 321548.279   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                           long  thrpt   40      295.719 ±     16.458  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                      long  thrpt   40       55.382 ±      1.462    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                  long  thrpt   40       19.332 ±     24.825  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm             long  thrpt   40        3.823 ±      5.112    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                     long  thrpt   40      231.758 ±    252.488  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm                long  thrpt   40       43.467 ±     47.391    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space              long  thrpt   40        6.600 ±     15.362  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space.norm         long  thrpt   40        1.472 ±      3.522    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                                long  thrpt   40       12.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                                 long  thrpt   40     1082.000                   ms
BenchmarkReferenceCountMap.benchmarkInserts                                          byte  thrpt   40  5879299.127 ± 376297.786   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                           byte  thrpt   40      309.390 ±     20.247  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                      byte  thrpt   40       55.367 ±      1.453    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                  byte  thrpt   40       19.146 ±     27.836  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm             byte  thrpt   40        4.195 ±      6.684    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                     byte  thrpt   40      231.390 ±    273.000  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm                byte  thrpt   40       47.017 ±     56.210    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space              byte  thrpt   40        8.092 ±     14.864  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space.norm         byte  thrpt   40        1.545 ±      2.832    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                                byte  thrpt   40        9.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                                 byte  thrpt   40     1025.000                   ms
```

Without this patch:
```
Benchmark                                                                     (arrayType)   Mode  Cnt        Score        Error   Units
BenchmarkReferenceCountMap.benchmarkInserts                                           int  thrpt   40  2918075.545 ± 282917.633   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                            int  thrpt   40      107.607 ±     12.005  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                       int  thrpt   40       39.248 ±      2.588    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                   int  thrpt   40        6.703 ±     11.156  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm              int  thrpt   40        4.784 ±      8.149    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                      int  thrpt   40        9.234 ±     15.844  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm                 int  thrpt   40        6.479 ±     12.085    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space               int  thrpt   40        0.329 ±      1.169  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space.norm          int  thrpt   40        0.149 ±      0.529    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                                 int  thrpt   40        5.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                                  int  thrpt   40     3937.000                   ms
BenchmarkReferenceCountMap.benchmarkInserts                                        double  thrpt   40  2992142.568 ± 308081.641   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                         double  thrpt   40      108.933 ±     13.055  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                    double  thrpt   40       39.257 ±      2.607    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                double  thrpt   40        7.656 ±     11.529  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm           double  thrpt   40        5.772 ±      9.056    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                   double  thrpt   40       15.783 ±     26.878  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm              double  thrpt   40       12.256 ±     24.282    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space            double  thrpt   40        0.128 ±      0.456  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space.norm       double  thrpt   40        0.059 ±      0.210    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                              double  thrpt   40        6.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                               double  thrpt   40     5888.000                   ms
BenchmarkReferenceCountMap.benchmarkInserts                                          long  thrpt   40  2870439.234 ± 361973.085   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                           long  thrpt   40      106.993 ±     13.969  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                      long  thrpt   40       39.459 ±      3.077    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                  long  thrpt   40        7.717 ±     13.777  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm             long  thrpt   40        8.231 ±     14.968    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                     long  thrpt   40        9.283 ±     18.181  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm                long  thrpt   40       10.991 ±     22.281    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                                long  thrpt   40        4.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                                 long  thrpt   40     3504.000                   ms
BenchmarkReferenceCountMap.benchmarkInserts                                          byte  thrpt   40  3037850.852 ± 340457.724   ops/s
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate                           byte  thrpt   40      111.818 ±     13.477  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.alloc.rate.norm                      byte  thrpt   40       39.368 ±      2.787    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space                  byte  thrpt   40        6.548 ±     11.220  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Eden_Space.norm             byte  thrpt   40        4.933 ±      8.394    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen                     byte  thrpt   40       17.228 ±     28.023  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Old_Gen.norm                byte  thrpt   40       13.105 ±     21.947    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space              byte  thrpt   40        0.279 ±      0.993  MB/sec
BenchmarkReferenceCountMap.benchmarkInserts:·gc.churn.G1_Survivor_Space.norm         byte  thrpt   40        0.175 ±      0.622    B/op
BenchmarkReferenceCountMap.benchmarkInserts:·gc.count                                byte  thrpt   40        5.000               counts
BenchmarkReferenceCountMap.benchmarkInserts:·gc.time                                 byte  thrpt   40     5816.000                   ms
```